pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "pillow>=11.1.0",
    "python-telegram-bot[job-queue]==20.7",
    "requests>=2.32.3",
    "telegram>=0.0.1",
    "trafilatura>=2.0.0",
    "twilio>=9.4.6",
    "yfinance>=0.2.54",
]
